Add lane to build wasm+threads version of icu (#173)
If you are building for wasm w/ pthreads enabled, you cannot link in a library that wasn't built w/ threads enabled. This PR adds another leg to produce a lib with threads support.

The emscripten docker container was also updated to 2.0.23 - the same version we use in dotnet/runtime.

NOTE: This change also patches emcmake.py within emscripten. This is to work around an issue with the version we are on where --experimental-wasm-threads is not passed to node and the build will end up failing. It's fixed in later versions.

The emscripten patched files are:

https://github.com/emscripten-core/emscripten/blob/b15ca40ca02f469de3bf9a4de67a2655aede8974/emcc.py#L613-L615

https://github.com/emscripten-core/emscripten/blob/b15ca40ca02f469de3bf9a4de67a2655aede8974/emcmake.py#L40